feat(mpc): port signed integer operations
Add signed MPC arithmetic, comparisons, conversions, shifts, and
checked overflow handling on top of upstream main while preserving
main's current contract layout.

Cover signed behavior across integer widths, including MIN / -1,
plain mul wraparound, signed shifts, and checked overflow reverts.

+++ contracts/mocks/utils/mpc/ValidateCiphertext256TestsContract.sol
@@ -51,7 +51,7 @@ contract ValidateCiphertext256TestsContract {
         gtUint256 validatedValue = MpcCore.validateCiphertext(input);
         
         // Add 1 to the value
-        gtUint256 one = MpcCore.setPublic256(1);
+        gtUint256 one = MpcCore.setPublic256(uint256(1));
         gtUint256 result = MpcCore.add(validatedValue, one);
         
         // Decrypt and return
